map out session window object

+++ db/migrate/20181005225229_create_condition_swells.rb
@@ -1,8 +1,8 @@
 class CreateConditionSwells < ActiveRecord::Migration[5.2]
   def change
     create_table :condition_swells do |t|
-      t.integer :min_height
-      t.integer :max_height
+      t.float :min_height
+      t.float :max_height
       t.integer :min_period
       t.string :direction, default: [], array: true
       t.timestamps
